+++ source/fitz/draw-scale-simple.c
@@ -414,11 +414,11 @@ check_weights(fz_weights *weights, int j, int w, float x, float wf)
 		weights->index[maxidx-1] += 256-sum;
 	/* Otherwise, if we are the first pixel, and it's fully covered, then
 	 * adjust it. */
-	else if ((j == 0) && (x < 0.0001F) && (sum != 256))
+	else if ((j == 0) && (x < 0.0001f) && (sum != 256))
 		weights->index[maxidx-1] += 256-sum;
 	/* Finally, if we are the last pixel, and it's fully covered, then
 	 * adjust it. */
-	else if ((j == w-1) && ((float)w-wf < 0.0001F) && (sum != 256))
+	else if ((j == w-1) && (w - wf < 0.0001f) && (sum != 256))
 		weights->index[maxidx-1] += 256-sum;
 }
 
@@ -1616,7 +1616,7 @@ fz_scale_pixmap_cached(fz_context *ctx, const fz_pixmap *src, float x, float y, 
 	else
 	{
 		dst_x_int = floorf(x);
-		x -= (float)dst_x_int;
+		x -= dst_x_int;
 		dst_w_int = (int)ceilf(x + w);
 	}
 	/* dst_y_int is calculated to be the top of the scaled image, and
@@ -1637,7 +1637,7 @@ fz_scale_pixmap_cached(fz_context *ctx, const fz_pixmap *src, float x, float y, 
 	else
 	{
 		dst_y_int = floorf(y);
-		y -= (float)dst_y_int;
+		y -= dst_y_int;
 		dst_h_int = (int)ceilf(y + h);
 	}
